refactor: split script.js into modules
