+++ scripts/clearKeyboard.js
@@ -1,0 +1,6 @@
+const clearKeyboard = () => {
+  const kbRows = document.querySelectorAll('.keyboard__row');
+  kbRows.forEach((el) => { const row = el; row.innerHTML = ''; });
+};
+
+export default clearKeyboard;
