+++ scripts/findCursorPos.js
@@ -1,0 +1,3 @@
+const findCursorPos = (rowLengths, cursorPos) => rowLengths.findIndex((el) => el[1] >= cursorPos);
+
+export default findCursorPos;
